Fix  not infering  (#1)
