+++ projects/fluig-dataset/package.json
@@ -1,6 +1,6 @@
 {
   "name": "fluig-dataset",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "license": "MIT",
   "description": "Serviço angular para consumo de datasets no Fluig",
   "author": "Gabriel Nascimento",
