+++ projects/fluig-dataset/src/public-api.ts
@@ -1,10 +1,7 @@
 /*
  * Public API Surface of fluig-dataset
  */
-export {
-  FLUIG_DATASET_CONFIG,
-  FluigDatasetConfig,
-} from './lib/fluig-dataset.config';
-
+export { FLUIG_DATASET_CONFIG } from './lib/fluig-dataset.config';
+export * from './lib/fluig-dataset';
 export * from './lib/fluig-dataset.service';
 export * from './lib/fluig-dataset.provider';
